Move nr_of_residues_in_total() to structure module
